refactor(web): extract UnavailableState and MissingProductState
The empty state used when software selection is not available was a local
component of PatternSelectionUnavailable, so nothing outside the software
module could reuse it. Move the presentational part to core/UnavailableState
and the missing product messaging (network hint plus link to the network
settings) to product/MissingProductState.

No visible change: the software page renders exactly the same states.
